+++ lib/network/online/end_points.dart
@@ -2,3 +2,4 @@ const LOGIN = '/login_instructor/';
 const SIGNUP = '/register_instructor/';
 const RESTPASSWORD = '/reset_password/';
 const SUGGESTEDCOURSES = '/course_by_top_rating/';
+const GETPROFILE = '/profile/';
